feat: LottoTicket 클래스에서 랜덤에 대한 의존도를 낮추기 위해서 DI 사용하도록 수정
기존에 LottoTicket 클래스에서 직접적으로 생성해서 사용하던 의존성 객체를
생성자를 통해 주입받도록 수정해서 코드 결합도(coupling)를 낮춘다

LottoNumberGenerator 클래스에 대한 단위 테스트 코드도 함께 수정한ㄷ
LottoTicket 클래스에 대한 단위 테스트 코드도 함께 수정한다
